Splitting data into separate data, reader and axes submodules
