+++ virtual_rainforest/core/data.py
@@ -33,59 +33,10 @@ Adding data to a Data instance
 
 The :meth:`~virtual_rainforest.core.data.Data.load_dataarray` method validates a
 provided :class:`~xarray.DataArray` using a configurable system of axis validation
-methods, before the data is added  to a :class:`~virtual_rainforest.core.data.Data`
-instance. These validators are used to check that particular signatures of dimensions
-and coordinates in the provided :class:`~xarray.DataArray` are congruent with the core
-configuration parameters.
-
-Core axes and validation system
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-A virtual rainforest simulation has a set of core axes - such as spatial, temporal and
-soil depth axes - which have dimensions and possibly coordinates that are defined in the
-model configuration.  Within a given axis - for example the 'Spatial' axis - there can
-be several methods to map the data in a provided array onto the axis. The method
-selection depends on matching the _signature_ of a provided array for a particular axis:
-this is a combination of the named dimensions and whether the dimensions are associated
-with coordinates.  As an example, an array with simple ``x`` and ``y`` dimensions and no
-coordinates could be mapped onto a square grid, assuming that the lengths of each
-dimension match.
-
-At present, the signature also includes only the spatial
-:attr:`~virtual_rainforest.core.grid.Grid.grid_type` used in the
-:class:`~virtual_rainforest.core.data.Data` instance, but this will likely be expanded
-to include configuration details for other core axes.
-
-The AXIS_VALIDATORS registry is a dictionary that provides sets of validator methods for
-different axes. The list of validators for a given axis is then a mapping of data
-signatures to the function that will handle that given signature. The individual
-validation functions should take an input that matches the signature, validate it,
-implement any internal standardisation and then return the valid, standardised data.
-
-The set of validators provided in the AXIS_VALIDATORS registry can be extended using the
-:func:`~virtual_rainforest.core.data.add_validator` decorator, which adds a new
-pairing of a signature and method to the appropriate named axis in the registry.
-
-So, for example:
-
-.. code-block:: python
-
-    @add_validator("spatial", (("x", "y"), ("x", "y"), ("square",)))
-
-This adds a spatial validator that will map a :class:`~xarray.DataArray` with ``x`` and
-``y`` coordinates (and hence implicitly 'x' and 'y' dimensions) onto a square grid.
-
-.. code-block:: python
-
-    @add_validator("spatial", (("cell_id",), (), ("any",)))
-
-This adds a spatial validator that will map a :class:`~xarray.DataArray` with the
-``cell_id`` dimension (but no ``cell_id`` coordinates) onto _any_ spatial grid type: the
-underlying ``cell_id``  attribute of the grid is defined for all grid.
-
-All spatial loader methods standardise the spatial structure of the input data to use a
-single ``cell_id`` spatial axis, which maps data onto the cell IDs used for indexing in
-the :class:`~virtual_rainforest.core.grid.Grid` instance for the simulation.
+methods (:mod:`~virtual_rainforest.core.axes`),  before the data is added  to a
+:class:`~virtual_rainforest.core.data.Data` instance. These validators are used to check
+that particular signatures of dimensions and coordinates in the provided
+:class:`~xarray.DataArray` are congruent with the core configuration parameters.
 
 Adding data from a file
 -----------------------
@@ -99,25 +50,6 @@ The general solution for programmatically adding data from a file is to:
 However, the :meth:`~virtual_rainforest.core.data.Data.load_from_file` method
 automatically loads data from known formats defined in the
 :attr:`~virtual_rainforest.core.data.FILE_FORMAT_REGISTRY`.
-
-The FILE_FORMAT_REGISTRY
-~~~~~~~~~~~~~~~~~~~~~~~~
-
-The :attr:`~virtual_rainforest.core.data.FILE_FORMAT_REGISTRY` is used to register the
-set of known file formats for use in
-:meth:`~virtual_rainforest.core.data.Data.load_from_file`. This registry is extendable,
-so that new functions that implement the approach above for a given file format can be
-added to those supported by :meth:`~virtual_rainforest.core.data.Data.load_from_file`.
-This is done using the :func:`~virtual_rainforest.core.data.register_file_format_loader`
-decorator, which needs to specify the file formats supported (as a tuple of file
-suffixes) and then decorate a function that returns a :class:`~xarray.DataArray`
-suitable for use in :meth:`~virtual_rainforest.core.data.Data.load_dataarray`. For
-example:
-
-.. code-block:: python
-
-    @register_file_format_loader(('.tif', '.tiff'))
-    def new_function_to_load_tif_data(...):
 
 Using a data configuration
 --------------------------
@@ -149,98 +81,14 @@ configurations cannot define repeated variable names.
 
 from collections import UserDict
 from pathlib import Path
-from typing import Any, Callable, Optional
+from typing import Any, Optional
 
-import numpy as np
-from xarray import DataArray, load_dataset
+from xarray import DataArray
 
 from virtual_rainforest.core.config import ConfigurationError
-from virtual_rainforest.core.grid import GRID_REGISTRY, Grid
+from virtual_rainforest.core.grid import Grid
 from virtual_rainforest.core.logger import LOGGER, log_and_raise
-
-FILE_FORMAT_REGISTRY: dict[str, Callable] = {}
-"""A registry for different file format loaders
-
-This dictionary maps a tuple of file format suffixes onto a function that allows the
-data to be loaded. That loader function should coerce the data into an xarray DataArray.
-
-Users can register their own functions to load from a particular file format using the
-:func:`~virtual_rainforest.core.data.register_file_format_loader` decorator. The
-function itself should have the following signature:
-
-    func(file: Path, file_var: str) -> DataArray
-"""
-
-
-def register_file_format_loader(file_types: tuple[str]) -> Callable:
-    """Adds a data loader function to the data loader registry.
-
-    This decorator is used to register a function that loads data from a given file type
-    and coerces it to a DataArray.
-
-    TODO - How to ensure that grids are registered before this decorator is used.
-
-    Args:
-        file_types: A tuple of strings giving the file type that the function will map
-            onto the Grid. The strings should match expected file suffixes for the file
-            type.
-    """
-
-    def decorator_file_format_loader(func: Callable) -> Callable:
-
-        # Ensure file_type is an iterable
-        if isinstance(file_types, str):
-            _file_types = (file_types,)
-        else:
-            _file_types = file_types
-
-        # Register the mapper function for each combination of grid type and file type
-        for this_ft in _file_types:
-
-            if this_ft in FILE_FORMAT_REGISTRY:
-                LOGGER.debug(
-                    "Replacing existing data loader function for %s",
-                    this_ft,
-                )
-            else:
-                LOGGER.debug(
-                    "Adding data loader function for %s",
-                    this_ft,
-                )
-
-            FILE_FORMAT_REGISTRY[this_ft] = func
-
-        return func
-
-    return decorator_file_format_loader
-
-
-@register_file_format_loader(file_types=(".nc",))
-def load_netcdf(file: Path, file_var_name: str) -> DataArray:
-    """Loads a DataArray from a NetCDF file.
-
-    Args:
-        file: A Path for a NetCDF file containing the variable to load.
-        file_var_name: A string providing the name of the variable in the file.
-    """
-
-    # Note that this deliberately doesn't contain any INFO logging messages to maintain
-    # a simple logging sequence - load_from_file, load_dataarray - without unnecessary
-    # logger noise about the specific format unless there is an exception.
-
-    # Try and load the provided file
-    try:
-        dataset = load_dataset(file)
-    except FileNotFoundError:
-        log_and_raise(f"Data file not found: {file}", FileNotFoundError)
-    except ValueError as err:
-        log_and_raise(f"Could not load data from {file}: {err}.", ValueError)
-
-    # Check if file var is in the dataset
-    if file_var_name not in dataset:
-        log_and_raise(f"Variable {file_var_name} not found in {file}", KeyError)
-
-    return dataset[file_var_name]
+from virtual_rainforest.core.readers import FILE_FORMAT_REGISTRY
 
 
 class Data(UserDict):
@@ -293,6 +141,9 @@ class Data(UserDict):
             key: The name to store the data under
             value: The DataArray to be stored
         """
+
+        # Import core.axis functions here to avoid circular imports
+        from virtual_rainforest.core.axes import AXIS_VALIDATORS, get_validator
 
         if not isinstance(value, DataArray):
             log_and_raise(
@@ -437,274 +288,3 @@ class Data(UserDict):
 
         if not clean_load:
             raise ConfigurationError("Data configuration did not load cleanly")
-
-
-AXIS_VALIDATORS: dict[str, dict[tuple, Callable]] = {}
-"""A registry for different axis validators
-
-This dictionary is keyed by the name of a particular core axis (e.g. 'spatial'), and the
-value for that key is then another dictionary that keys a particular data array
-signature to a function that validates data with that signature.
-
-Users can register their own functions to map data onto a core axis using the
-:func:`~virtual_rainforest.core.data.register_axis_validator` decorator. The
-function itself should have the following signature:
-
-    func(data: Data, darray: DataArray) -> DataArray
-"""
-
-
-def register_axis_validator(axis: str, signature: tuple) -> Callable:
-    """Add a new validator and data signature to a core axis.
-
-    This decorator adds the decorated method to the AXIS_VALIDATORS registry to extend
-    or amend the available routines used to load and validate an array containing
-    spatial information. The decorator adds the method to the list of validators
-    provided for the given axis class attribute under the given signature.
-
-    The signature consists of a tuple containing:
-
-    * a tuple of dimension names required in the data array to use the function,
-    * a tuple of coordinate names required in the data array to use the function - this
-      can be an empty tuple if no coordinate names are required,
-    * a tuple of grid types that the loader can be used with. The "any" grid type
-      can be used to indicate that a loader will work with any grid type.
-
-    Args:
-        axis: The core axis that the validator applies to
-        signature: A tuple of the validator signature.
-    """
-
-    def decorator(func: Callable) -> Callable:
-
-        # Check the grid types are valid
-        grid_types = signature[2]
-        for gtype in grid_types:
-            if (gtype not in GRID_REGISTRY) and (gtype != "any"):
-                log_and_raise(
-                    f"Unknown grid type '{gtype}' decorating {func.__name__}",
-                    ValueError,
-                )
-
-        # Does the axis exist yet?
-        if axis not in AXIS_VALIDATORS:
-            AXIS_VALIDATORS[axis] = dict()
-
-        # Register the signature -> function map under the given axis, checking if the
-        # signature already exists
-        if signature in AXIS_VALIDATORS[axis]:
-            LOGGER.debug("Replacing existing %s validator: %s", axis, func.__name__)
-        else:
-            LOGGER.debug("Adding %s validator: %s", axis, func.__name__)
-
-        # Add the validator
-        AXIS_VALIDATORS[axis][signature] = func
-
-        return func
-
-    return decorator
-
-
-def get_validator(axis: str, data: Data, darray: DataArray) -> Optional[Callable]:
-    """Get the matching validator function for a data array on a given axis.
-
-    This function iterates over the registered validator functions for a given core axis
-    in the AXIS_VALIDATORS registry and matches the data array signatures against the
-    provided signatures. It returns either the appropriate function or None.
-
-    Args:
-        axis: The core axis to get a validator for
-        data: A Data object providing parameter information
-        darray: The data to match to a validator signature
-    """
-
-    # Check the axis exists
-    if axis not in AXIS_VALIDATORS:
-        raise ValueError(f"Unknown core axis: {axis}")
-
-    # Identify the correct validator routine from the data array signature
-    da_dims = set(darray.dims)
-    da_coords = set(darray.coords.variables)
-
-    # Loop over the available validators loooking for a congruent signature
-    # - dims _cannot_ be empty on a DataArray (they default to dim_N strings) so are
-    #   guaranteed to be non-empty and must then match.
-    # - coords _can_ be empty: they just associate values with indices along the
-    #   dimension. So, the signature should match _specified_ coords but not the
-    #   empty set.
-    for (ld_dims, ld_coords, ld_grid_type), ld_fnm in AXIS_VALIDATORS[axis].items():
-
-        if (
-            set(ld_dims).issubset(da_dims)
-            and set(ld_coords).issubset(da_coords)
-            and ((data.grid.grid_type in ld_grid_type) or ("any" in ld_grid_type))
-        ):
-
-            # Retrieve the method associated with the loader signature from the Data
-            # object.
-            return ld_fnm
-
-    return None
-
-
-@register_axis_validator("spatial", (("cell_id",), ("cell_id",), ("any",)))
-def spld_cellid_coord_any(data: Data, darray: DataArray) -> DataArray:
-    """Spatial loader for cell id coordinates onto any grid.
-
-    In this loader, the DataArray has a cell_id dimension with valued coordinates, which
-    should map onto the grid cell ids, allowing for a subset of ids. Because this method
-    simply maps data to grid cells by id, it should apply to _any_ arbitrary grid setup.
-
-    Args:
-        data: A Data instance used to access validation information
-        darray: A data array containing spatial information to be validated
-
-    Returns:
-        A validated dataarray with a single cell id spatial dimension
-    """
-
-    da_cell_ids = darray["cell_id"].values
-
-    if len(np.unique(da_cell_ids)) != len(da_cell_ids):
-        raise ValueError("The data cell ids contain duplicate values.")
-
-    if not set(data.grid.cell_id).issubset(da_cell_ids):
-        raise ValueError("The data cell ids are not a superset of grid cell ids.")
-
-    # Now ensure sorting and any subsetting:
-    # https://stackoverflow.com/questions/8251541
-    da_sortorder = np.argsort(da_cell_ids)
-    gridid_pos = np.searchsorted(da_cell_ids[da_sortorder], data.grid.cell_id)
-    da_indices = da_sortorder[gridid_pos]
-
-    return darray.isel(cell_id=da_indices)
-
-
-@register_axis_validator("spatial", (("cell_id",), (), ("any",)))
-def spld_cellid_dim_any(data: Data, darray: DataArray) -> DataArray:
-    """Spatial loader for cell id dimension onto any grid.
-
-    In this loader, the DataArray only has a cell_id dimension so assumes that the
-    values are provided in the same sequence as the grid cell ids. Because this method
-    simply maps data to grid cells by id, it should apply to _any_ arbitrary grid setup.
-
-    Args:
-        data: A Data instance used to access validation information
-        darray: A data array containing spatial information to be validated
-
-    Returns:
-        A validated dataarray with a single cell id spatial dimension
-
-    """
-
-    # Cell ID is only a dimenson with a give length - assume the order correct
-    # and check the right number of cells found
-    n_found = darray["cell_id"].size
-    if data.grid.n_cells != n_found:
-        raise ValueError(
-            f"Grid defines {data.grid.n_cells} cells, data provides {n_found}"
-        )
-
-    return darray
-
-
-@register_axis_validator("spatial", (("x", "y"), ("x", "y"), ("square",)))
-def spld_xy_coord_square(data: Data, darray: DataArray) -> DataArray:
-    """Spatial loader for XY coordinates onto a square grid.
-
-    In this loader, the DataArray has a x and y dimensions with valued coordinates,
-    which should map onto the grid cell ids, allowing for a subset of ids.
-
-    Args:
-        data: A Data instance used to access validation information
-        darray: A data array containing spatial information to be validated
-
-    Returns:
-        A validated dataarray with a single cell id spatial dimension
-    """
-
-    # Get x and y coords to check the extents and cell coverage.
-    #
-    # TODO - Note that mapping all the cells here is a bit extreme with a square grid -
-    # could just map one row and column to confirm the indexing into the data, but this
-    # is a more general solution.
-
-    # Use .stack() to convert the axis into stacked pairs of values
-    xy_pairs = darray.stack(cell=("y", "x")).coords
-    idx_pairs = darray.drop_vars(("x", "y")).stack(cell=("y", "x")).coords
-
-    # Get the mapping of points onto the grid
-    idx_x, idx_y = data.grid.map_xy_to_cell_indexing(
-        x_coords=xy_pairs["x"].values,
-        y_coords=xy_pairs["y"].values,
-        x_idx=idx_pairs["x"].values,
-        y_idx=idx_pairs["y"].values,
-        strict=False,
-    )
-
-    # TODO - fine a way to enable strict = True - probably just kwargs.
-
-    # Now remap the grids from xy to cell_id - this uses the rather under
-    # described vectorized indexing feature in xarray:
-    #
-    #   https://docs.xarray.dev/en/stable/user-guide/indexing.html
-    #
-    # Specifically using DataArray.isel to select indexed dimensions by name, which
-    # avoids issues with different permutations of the axes - and map XY onto a common
-    # cell_id
-
-    return darray.isel(
-        x=DataArray(idx_x, dims=["cell_id"]), y=DataArray(idx_y, dims=["cell_id"])
-    )
-
-
-@register_axis_validator("spatial", (("x", "y"), (), ("square",)))
-def spld_xy_dim_square(data: Data, darray: DataArray) -> DataArray:
-    """Spatial loader for XY dimensions onto a square grid.
-
-    In this loader, the DataArray has x and y dimension but no coordinates along those
-    dimensions. The assumption here is then that those spatial axes must describe the
-    same array shape as the square grid.
-
-    Args:
-        data: A Data instance used to access validation information
-        darray: A data array containing spatial information to be validated
-
-    Returns:
-        A validated dataarray with a single cell id spatial dimension
-    """
-
-    # Otherwise the data array must be the same shape as the grid
-    if data.grid.cell_nx != darray.sizes["x"] or data.grid.cell_ny != darray.sizes["y"]:
-        raise ValueError("Data XY dimensions do not match square grid")
-
-    # Use DataArray.stack to combine the x and y into a multiindex called cell_id, with
-    # x varying fastest (cell_id goes from top left to top right, then down by rows),
-    # and then use these stacked indices to map the 2D onto grid cell order, using
-    # isel() to avoid issues with dimension ordering.
-    darray_stack = darray.stack(cell_id=("y", "x"))
-
-    return darray.isel(
-        x=DataArray(darray_stack.coords["x"].values, dims=["cell_id"]),
-        y=DataArray(darray_stack.coords["y"].values, dims=["cell_id"]),
-    )
-
-
-class DataGenerator:
-    """Generate artificial data.
-
-    Currently just a signature sketch.
-    """
-
-    def __init__(
-        self,
-        # grid: GRID,
-        spatial_axis: str,
-        temporal_axis: str,
-        temporal_interpolation: np.timedelta64,
-        seed: Optional[int],
-        method: str,  # one of the numpy.random.Generator methods
-        **kwargs: Any,
-    ) -> None:
-
-        pass
